php ok - js ko - css ko

+++ deleteUser.php
@@ -1,10 +1,17 @@
 <?php
-    require ('config/connect.php');
+session_start();
+
+if ($_SESSION['isAdmin'] == true) {
+    require_once('config/connect.php');
 
     $id = htmlspecialchars($_GET['id']);
 
     $req = $bdd->prepare('DELETE FROM user WHERE id= :id');
-    $req->execute(array('id'=>$id));
-    header('Location: userAdmin.php');
+    $req->execute(array('id' => $id));
+    header('Location:' . $_SERVER['HTTP_REFERER']);
     $req->CloseCursor();
+} else {
+    echo "vous devez être connecté pour soumettre ce formulaire - rien n'a été modifié";
+}
+
 ?>
